+++ apps/ui/src/vite-env.d.ts
@@ -1,4 +1,1 @@
 /// <reference types="vite/client" />
-interface ImportMetaEnv {
-  VITE_BILBOMD_BACKEND_PORT: string
-}